Move helm_release specialized conversion into pkg/convert/helm.go
